Drive straight with BrainInertial
This is a program that I converted from Blocks to Python in order to be able to add it into our program eventually. It is a function to drive straight without drift for autonomous skills with Brain Inertial. As long as we name the motors to LeftMotor and RightMotor, it should work and it doesn't need to have the same ports.